Add support for calculating a delta file
The new SignatureLookup class loads a signature and allows lookup up
blocks that have a matching rolling checksum. The DeltaGenerator class
takes data from the new file and uses the DeltaCommandGenerator to
create the delta file.

Tests have been added for all the new operations, although they are not
really thorough (they don't test huge files or block matches at various
offsets in a file).

The generated delta file should be compatible with the C librsync (the
C librsync doesn't implement an optimization of literal commands that
the format allows).